test: add integration tests with Azurite emulator (#1202)
* feat: added integration test with Azurite emulator and a feature to configure the internal TableServiceClient
Added integration tests with Azurite emulator to test the working of the library. Added a feature to configure the TableServiceClient as part of the AzureTableStorageModule.forRoot and forRootAsync options, to configure allowInsecureConnection needed to connect to the local emulator.

* test: suppress node warnings to fix tests

The ci tests using node 21 throw a deprecated warning (DeprecationWarning: The `punycode` module is deprecated. Please use a userland alternative instead.) that prevents the tests from running. punycode is a dependency of azurite.

* style: remove logging

Removed logging from azure-table.service.ts

* ci(test): configured jest with --runInBand

To run the integration tests within a medium resource class, the tests must run serially instead of in parallel.